refactor packages, add rest controller
