--- src/application.properties
@@ -1,8 +1,0 @@
-server.port=8081
-
-spring.thymeleaf.cache=false
-spring.thymeleaf.enabled=true
-spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html
-
-spring.application.name=Bootstrap Spring Boot
