+++ src/main/java/com/example/DemoApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
