+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.application.name=Bootstrap Spring Boot
+server.port=8080
