+++ src/test/java/com/example/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
